Made 'plugins' path full instead of relative
Bug occurred when running Hindsight from a different directory than the script resided in; eg: > python.exe c:\tools\hindsight\hindsight.py.
Hindsight would search for a directory called 'plugins' in the directory from which the terminal was, and would fail to find it.  Reworked the plugins import code to use the directory the script is in, not the working directory.