Log to current working directory by default.
This strikes me as more typical but my motivation is packaging. I think
any package manager is going to have a problem with logging to the same
directory as the binary because in a global installation users wont have
write permission to that directory. In nixpkgs it's even more
problematic because packages are installed to a read-only filesystem.