Pass no_copy to Chrome instances under test.
One of the reasons the tests were failing is that the temporary
directories weren't passed in and therefore didn't exist. Better to
forgo creating them in the first place.

Also call the get_downloads method from the instance rather than passing
the instance in, which is a more conventional python pattern.